Item13989: Make guest sessions a bit more persistent
If a user gets a guest session by visiting a page that requires
sessions, allow it to become persistent.  This means we don't need
SwitchBoard to set the sessionRequired context.

The cleanup code needs to be moved into the logout processing, so that
sessions are deleted when logging out.

Also fix a bug in Registration.   Register called
LoginManager::userLoggedIn() to cause the session to switch to the newly
registered user,  but this was not sticking.  userLoggedIn needs to add
the newly created cookie into the response.