Item13957: Item13958: Fix white space issues in : indents
Empty indent lines were rendered as empty divs, which causes them
to disappear when viewed.  Insert a &nbsp; into empyty div, so they
result in a blank line.

The Wysiwyg editor eats blank lines that are inserted between two
indented blocks.  The TML2HTML failed to recognize that it was in a
paragraph and should close the paragraph when encountering a blank line.

Note, this causes insertion of a blank line by wysiwyg:

   : Indent
      : Indent2
Un-indented text

ends up with a blank line before the un-indented text. I think that this
is acceptable,  I could not figure out how to prevent it in this case,
while preserving a blank line between two indent items.   NEEDS REVIEW!